// backoffice/src/app/shared/Model/user.ts
export class User {
  id: any;
  username: any;
  email: any;
  password: any;
  name: any;
  lastname: any;
  dateBirth: any;
  gender: any;
  job: any;
  adress: any;
  profilePic: any;
  identitycardPic: any;
  signaturePic: any;
  creationDate: any;
  active: any;
  verified: any;
  token: any;
  tokenCreationDate: any;
  verificationCode: any;
  badge: any;
}